+++ Cargo.toml
@@ -6,6 +6,4 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-galois_field = "0.1.11"
-gf = "0.6.1"
 hex = "0.4.3"
